+++ components/Contact/Contact.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import styles from "./Contact.module.css";
-import HyperLink from "../hyperLink";
+import HyperLink from "../hyperLink/";
 import ContactForm from "../ContactForm/";
 
 function Contact({links}) {
